Reorganise meta-structure of diss and header
In terms of diss.t, the main difference is that the new commands are
taken out and moved to header, for more consistency. Possibly other
minor alterations but I don't think so.

Header has a lot more changed. The redundant geometry line is taken
out, and the paragraph width specified in the shortguide is implemented.
Duplicate packages are removed also. Otherwise, the main changes are
just tidying up and reformatting things, as well as the addition of the
custom commands imported from diss.t.

Changes to be committed:
      modified:   dissertation.tex
      modified:   header.tex