Start implementing feedback from Nadia
For changes to tex, only really changed the [htbp] priorities of some
figure environments that were accidentally set to [htpb].

Otherwise, as title. Started from the beginning of the printed out draft
with annotated feedback from Nadia implementing the suggested changes.
Got as far as the first two subsections of section 1 in this commit.
Definitely have not finished with them yet though. See the diff for all
changes.