Multiple changes, finishing a lipsum free draft
Structurally, moved the addition diagram to the figures folder as I
should have done before. Also added a new diagram, with R code in a
comment.

Much of the changes are writing very basic beginnings for the lipsum
sections. I can't remember what a lot of them are though so just check
the diff. Good god, it's not like I'll be checking back on this anytime
soon. Hand in is on friday. It's crunch time. No time to be
version-controlling like a maniac.

Modern Family, chess and Skyrim 4 lyf.